fix: extracting the common code to shared and mock code in quickaction.js
